all  data

+++ mindspeed_mm/training.py
@@ -138,24 +138,26 @@ def pretrain(
     if one_logger:
         one_logger.log_metrics({"train_iterations_warmup": 5})
 
-    # NOTE For group data, we need to set initial_global_step_for_sampler
-    group_data = getattr(args.mm.data.dataloader_param, 'group_data', False)
 
-    if group_data:
-        # group sampler
-        timers("global-step-for-sampler-setup", log_level=0).start(barrier=True)
-        print_rank_0("use group sampler...")
-        global_step_for_sampler_txt = os.path.join(args.save, 'global_step_for_sampler.txt')
-        if os.path.exists(global_step_for_sampler_txt):
-            with open(global_step_for_sampler_txt, 'r') as f:
-                global_step_for_sampler = int(f.read().strip())
-        elif args.mm.data.dataloader_param.initial_global_step_for_sampler != 0:
-            global_step_for_sampler = args.mm.data.dataloader_param.initial_global_step_for_sampler
-        else:
-            global_step_for_sampler = 0
-        setattr(args.mm.data.dataloader_param, 'initial_global_step_for_sampler', global_step_for_sampler)
-        print_rank_0(f"global_step_for_sampler: {args.mm.data.dataloader_param.initial_global_step_for_sampler}")
-        timers("global-step-for-sampler-setup").stop()
+    if torch.distributed.get_rank() == 0:
+        # NOTE For group data, we need to set initial_global_step_for_sampler
+        group_data = getattr(args.mm.data.dataloader_param, 'group_data', False)
+        if group_data:
+            # group sampler
+            timers("global-step-for-sampler-setup", log_level=0).start()
+            print_rank_0("use group sampler...")
+            global_step_for_sampler_txt = os.path.join(args.save, 'global_step_for_sampler.txt')
+            if os.path.exists(global_step_for_sampler_txt):
+                with open(global_step_for_sampler_txt, 'r') as f:
+                    global_step_for_sampler = int(f.read().strip())
+            elif args.mm.data.dataloader_param.initial_global_step_for_sampler != 0:
+                global_step_for_sampler = args.mm.data.dataloader_param.initial_global_step_for_sampler
+            else:
+                global_step_for_sampler = 0
+            setattr(args.mm.data.dataloader_param, 'initial_global_step_for_sampler', global_step_for_sampler)
+            print_rank_0(f"global_step_for_sampler: {args.mm.data.dataloader_param.initial_global_step_for_sampler}")
+            timers("global-step-for-sampler-setup").stop()
+    torch.distributed.barrier()
         
     # Model, optimizer, and learning rate.
     timers("model-and-optimizer-setup", log_level=0).start(barrier=True)
@@ -236,20 +238,21 @@ def pretrain(
                 num_floating_point_operations_so_far,
             )
 
-        # NOTE For group data, we need to save the global step for sampler.
-        group_data = getattr(args.mm.data.dataloader_param, 'group_data', False)
-
-        if group_data and not extreme_error_flag:
-            # group sampler
-            timers("global-step-for-sampler-txt-delete-setup", log_level=0).start(barrier=True)
-            global_step_for_sampler_txt = os.path.join(args.save, 'global_step_for_sampler.txt')
-            if os.path.exists(global_step_for_sampler_txt):
-                print_rank_0("delete global_step_for_sampler.txt...")
-                with open(global_step_for_sampler_txt, 'r') as f:
-                    global_step_for_sampler = int(f.read().strip())
-                print_rank_0(f"global_step_for_sampler: {global_step_for_sampler}, and we will reset it to 0...")
-                os.remove(global_step_for_sampler_txt)
-            timers("global-step-for-sampler-txt-delete-setup").stop()
+        if torch.distributed.get_rank() == 0:
+            # NOTE For group data, we need to save the global step for sampler.
+            group_data = getattr(args.mm.data.dataloader_param, 'group_data', False)
+            if group_data and not extreme_error_flag:
+                # group sampler
+                timers("global-step-for-sampler-txt-delete-setup", log_level=0).start()
+                global_step_for_sampler_txt = os.path.join(args.save, 'global_step_for_sampler.txt')
+                if os.path.exists(global_step_for_sampler_txt):
+                    print_rank_0("delete global_step_for_sampler.txt...")
+                    with open(global_step_for_sampler_txt, 'r') as f:
+                        global_step_for_sampler = int(f.read().strip())
+                    print_rank_0(f"global_step_for_sampler: {global_step_for_sampler}, and we will reset it to 0...")
+                    os.remove(global_step_for_sampler_txt)
+                timers("global-step-for-sampler-txt-delete-setup").stop()
+        torch.distributed.barrier()
     else:
         print_rank_0("skipping training (--skip-train is on) ...")
 
@@ -487,24 +490,24 @@ def train(
             params_norm,
             num_zeros_in_grad,
         )
-
-        if extreme_error_flag: 
-            # NOTE For group data, we need to save the global step for sampler.
-            group_data = getattr(args.mm.data.dataloader_param, 'group_data', False)
-            if group_data:
-                # group sampler
-                timers("global-step-for-sampler-txt-save-setup", log_level=0).start(barrier=True)
-                print_rank_0("save global_step_for_sampler.txt")
-                initial_global_step_for_sampler = getattr(args.mm.data.dataloader_param, 'initial_global_step_for_sampler', 0)
-                global_step_for_sampler_txt = os.path.join(args.save, 'global_step_for_sampler.txt')
-                global_step_for_sampler = initial_global_step_for_sampler + iteration - initial_iteration
-                print_rank_0(f"initial_global_step_for_sampler: {initial_global_step_for_sampler}")
-                print_rank_0(f"current_global_step_for_sampler: {global_step_for_sampler}")
-                with open(global_step_for_sampler_txt, 'w') as f:
-                    f.write(str(global_step_for_sampler))
-                timers("global-step-for-sampler-txt-save-setup").stop()
-            break
-
+        if torch.distributed.get_rank() == 0:
+            if extreme_error_flag: 
+                # NOTE For group data, we need to save the global step for sampler.
+                group_data = getattr(args.mm.data.dataloader_param, 'group_data', False)
+                if group_data:
+                    # group sampler
+                    timers("global-step-for-sampler-txt-save-setup", log_level=0).start()
+                    print_rank_0("save global_step_for_sampler.txt")
+                    initial_global_step_for_sampler = getattr(args.mm.data.dataloader_param, 'initial_global_step_for_sampler', 0)
+                    global_step_for_sampler_txt = os.path.join(args.save, 'global_step_for_sampler.txt')
+                    global_step_for_sampler = initial_global_step_for_sampler + iteration - initial_iteration
+                    print_rank_0(f"initial_global_step_for_sampler: {initial_global_step_for_sampler}")
+                    print_rank_0(f"current_global_step_for_sampler: {global_step_for_sampler}")
+                    with open(global_step_for_sampler_txt, 'w') as f:
+                        f.write(str(global_step_for_sampler))
+                    timers("global-step-for-sampler-txt-save-setup").stop()
+                break
+        torch.distributed.barrier()
         # Autoresume
         if args.adlr_autoresume and (iteration % args.adlr_autoresume_interval == 0):
             check_adlr_autoresume_termination(
@@ -569,21 +572,22 @@ def train(
             )
             saved_checkpoint = True
 
-            # NOTE For group data, we need to save the global step for sampler.
-            group_data = getattr(args.mm.data.dataloader_param, 'group_data', False)
-
-            if group_data:
-                # group sampler
-                timers("global-step-for-sampler-txt-save-setup", log_level=0).start(barrier=True)
-                print_rank_0("save global_step_for_sampler.txt")
-                initial_global_step_for_sampler = getattr(args.mm.data.dataloader_param, 'initial_global_step_for_sampler', 0)
-                global_step_for_sampler_txt = os.path.join(args.save, 'global_step_for_sampler.txt')
-                global_step_for_sampler = initial_global_step_for_sampler + iteration - initial_iteration
-                print_rank_0(f"initial_global_step_for_sampler: {initial_global_step_for_sampler}")
-                print_rank_0(f"current_global_step_for_sampler: {global_step_for_sampler}")
-                with open(global_step_for_sampler_txt, 'w') as f:
-                    f.write(str(global_step_for_sampler))
-                timers("global-step-for-sampler-txt-save-setup").stop()
+            if torch.distributed.get_rank() == 0:
+                # NOTE For group data, we need to save the global step for sampler.
+                group_data = getattr(args.mm.data.dataloader_param, 'group_data', False)
+                if group_data:
+                    # group sampler
+                    timers("global-step-for-sampler-txt-save-setup", log_level=0).start()
+                    print_rank_0("save global_step_for_sampler.txt")
+                    initial_global_step_for_sampler = getattr(args.mm.data.dataloader_param, 'initial_global_step_for_sampler', 0)
+                    global_step_for_sampler_txt = os.path.join(args.save, 'global_step_for_sampler.txt')
+                    global_step_for_sampler = initial_global_step_for_sampler + iteration - initial_iteration
+                    print_rank_0(f"initial_global_step_for_sampler: {initial_global_step_for_sampler}")
+                    print_rank_0(f"current_global_step_for_sampler: {global_step_for_sampler}")
+                    with open(global_step_for_sampler_txt, 'w') as f:
+                        f.write(str(global_step_for_sampler))
+                    timers("global-step-for-sampler-txt-save-setup").stop()
+            torch.distributed.barrier()
 
         # Exiting based on duration
         if args.exit_duration_in_mins:
